fix: tempfile prefix windows support
